fix: correcciión de bug de calculo de ganancia

+++ src/main/java/cl/javadevs/springsecurityjwt/dtos/reserva/response/DtoReporteResponse.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class DtoReporteResponse {
 
     private String servicioNombre;
-    private Integer montoTotal;
+    private Long montoTotal;
     private Integer cantidadReservas;
 }
